requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
djangorestframework==3.9.4
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.1
pkg-resources==0.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.9
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
